d/rules: Explicitly disable use of libraries we don't build-depend on
SDL doesn't build-depend on JACK or the proprietary bcm_host and Vivante
graphics APIs, so these would not get enabled in official Debian builds.
For consistency, if they happen to be installed on a system where a
contributor is doing their own local build, don't enable them
there either.

In particular, this ensures that the old bcm_host driver for Raspberry Pi
is not enabled. This is an outdated kernel driver which has been
superseded by KMS/DRM.

Closes: #1012863